requirements.txt
langchain-huggingface
langchain-core
langchain-text-splitters
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
openai>=1.54.0
python-multipart>=0.0.12
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
pypdf>=5.0.0
tiktoken>=0.8.0
requests>=2.32.0